+++ x/evm/keeper/funtoken_from_coin_test.go
@@ -225,8 +225,6 @@ func (s *FunTokenFromCoinSuite) TestConvertCoinToEvmAndBack() {
 	)
 	s.Require().ErrorContains(err, "insufficient funds")
 
-	deps.ResetGasMeter()
-
 	s.T().Log("Convert erc-20 to back to bank coin")
 	_, err = deps.EvmKeeper.CallContract(
 		deps.Ctx,
@@ -254,8 +252,6 @@ func (s *FunTokenFromCoinSuite) TestConvertCoinToEvmAndBack() {
 	balance, err = deps.EvmKeeper.ERC20().BalanceOf(funToken.Erc20Addr.Address, alice.EthAddr, deps.Ctx)
 	s.Require().NoError(err)
 	s.Require().Equal("0", balance.String())
-
-	deps.ResetGasMeter()
 
 	s.T().Log("sad: Convert more erc-20 to back to bank coin, insufficient funds")
 	_, err = deps.EvmKeeper.CallContract(
@@ -358,7 +354,6 @@ func (s *FunTokenFromCoinSuite) TestNativeSendThenPrecompileSend() {
 	}.Assert(s.T(), deps)
 
 	s.T().Log("call test contract")
-	deps.ResetGasMeter()
 	newSendAmtSendToBank := new(big.Int).Quo(sendAmt, big.NewInt(2))
 	newSendAmtEvmTransfer := evm.NativeToWei(newSendAmtSendToBank)
 	evmResp, err := deps.EvmKeeper.CallContract(
@@ -402,7 +397,6 @@ func (s *FunTokenFromCoinSuite) TestNativeSendThenPrecompileSend() {
 		BalanceERC20: big.NewInt(0),
 	}.Assert(s.T(), deps)
 
-	deps.ResetGasMeter()
 	evmResp, err = deps.EvmKeeper.CallContract(
 		deps.Ctx,
 		embeds.SmartContract_TestNativeSendThenPrecompileSendJson.ABI,